% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moviemeter.r
\name{mm_get_movie_info}
\alias{mm_get_movie_info}
\title{Retrieve film information}
\usage{
mm_get_movie_info(movie_id)
}
\arguments{
\item{movie_id}{either a MovieMeter ID or IMDB code}
}
\value{
\code{list}
}
\description{
Pass in either a MovieMeter ID or IMDB code to retrieve movie info.
}
\details{
You need a MovieMeter API KEY. See \link{moviemeter_api_key} for more info.
}
\note{
MovieMeter requests citations when publishing any derivative work from their data.
}
\examples{
mm_get_movie_info(5208)
}
\references{
\url{http://wiki.moviemeter.nl/index.php/API}
}

