% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moviemeter.r
\name{mm_search}
\alias{mm_search}
\title{Search for films}
\usage{
mm_search(terms)
}
\arguments{
\item{terms}{atomic character vector to use in the search}
}
\value{
\code{tibble}
}
\description{
You need a MovieMeter API KEY. See \link{moviemeter_api_key} for more info.
}
\note{
MovieMeter requests citations when publishing any derivative work from their data.
}
\examples{
mm_search("brave little toaster")
}
\references{
\url{http://wiki.moviemeter.nl/index.php/API}
}

