% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moviemeter-package.R
\docType{package}
\name{moviemeter}
\alias{moviemeter}
\alias{moviemeter-package}
\title{Tools to work with the MovieMeter API}
\description{
Get an API key: \url{https://www.moviemeter.nl/site/registerclient/}
}
\details{
Read about their API \url{http://wiki.moviemeter.nl/index.php/API}
}
\author{
Bob Rudis (bob@rud.is)
}

