% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.r
\name{moviemeter_api_key}
\alias{moviemeter_api_key}
\title{Get or set MOVIEMETER_API_KEY value}
\usage{
moviemeter_api_key(force = FALSE)
}
\arguments{
\item{force}{force setting a new MovieMeter API key for the current environment?}
}
\value{
atomic character vector containing the MovieMeter API key
}
\description{
The API wrapper functions in this package all rely on a MovieMeter API
key residing in the environment variable \code{MOVIEMETER_API_KEY}. The
easiest way to accomplish this is to set it in the `\code{.Renviron}` file in your
home directory.
}

